applications: asset_tracker: Add user interface module
This patch adds a user interface (UI) module to handle buttons
and LEDs. The application can provide an event handler and
receive events from the UI module on button state changes.
It has two modes of LED operation:
- PWM-based for RGB
- Non-PWM for simple signalling of application state.

The intention is to expand the module to also handle pin
control, light sensor and buzzer.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>